+++ src/VsChromium/Features/ToolWindows/SourceExplorer/CancelSearchCommandHandler.cs
@@ -17,7 +17,7 @@ namespace VsChromium.Features.ToolWindows.SourceExplorer {
 
     public override CommandID CommandId {
       get {
-        return new CommandID(GuidList.GuidVsChromiumCmdSet, PkgCmdIdList.CmdidCancelSearch);
+        return new CommandID(GuidList.GuidVsChromiumCmdSet, (int)PkgCmdIdList.CmdidCancelSearch);
       }
     }
