+++ src/VsChromium/Features/ToolWindows/SourceExplorer/SearchTextCommandHandler.cs
@@ -20,7 +20,7 @@ namespace VsChromium.Features.ToolWindows.SourceExplorer {
 
     public override CommandID CommandId {
       get {
-        return new CommandID(GuidList.GuidVsChromiumCmdSet, PkgCmdIdList.CmdidSearchText);
+        return new CommandID(GuidList.GuidVsChromiumCmdSet, (int)PkgCmdIdList.CmdidSearchText);
       }
     }
